Return CommitFilesAsync back to IGitRepo
