+++ test/Microsoft.DotNet.Darc.Tests/DependencyFileManagerTests.cs
@@ -65,7 +65,7 @@ public class DependencyFileManagerTests
         "https://pkgs.dev.azure.com/dnceng/public/_packaging/darc-pub-dotnet-corefx-4ac4c036/nuget/v3/index.json" })]
     public async Task UpdatePackageSourcesTests(string testName, string[] managedFeeds)
     {
-        DependencyFileManager dependencyFileManager = new DependencyFileManager(null, null, new VersionDetailsParser(), NullLogger.Instance);
+        DependencyFileManager dependencyFileManager = new DependencyFileManager(null, new VersionDetailsParser(), NullLogger.Instance);
 
         string inputNugetPath = Path.Combine(
             Environment.CurrentDirectory,
@@ -118,7 +118,7 @@ public class DependencyFileManagerTests
     [TestCase("NoDuplicatedDifferentConditions.props", false)]
     public void VerifyNoDuplicatedPropertiesTests(string inputFileName, bool hasDuplicatedProps)
     {
-        DependencyFileManager dependencyFileManager = new DependencyFileManager(null, null, new VersionDetailsParser(), NullLogger.Instance);
+        DependencyFileManager dependencyFileManager = new DependencyFileManager(null, new VersionDetailsParser(), NullLogger.Instance);
 
         string inputVersionPropsPath = Path.Combine(
             Environment.CurrentDirectory,
